Ticket #903: fix minor compilation warning with GCC 9.3.0

+++ tests/ClientServer/unit_tests/helpers/check_logger.c
@@ -494,8 +494,8 @@ START_TEST(test_logger_user)
     SOPC_Log_Trace(userLog, SOPC_LOG_LEVEL_WARNING, userLongLine3);
     SOPC_Free(userLongLine3);
     ck_assert(SOPC_Check_Logger_userLogCalled);
-    ck_assert_msg(strlen(SOPC_Check_Logger_lastUserLog) == SOPC_LOG_MAX_USER_LINE_LENGTH,
-                  "Was expecting len=%zu, but found %zu", SOPC_LOG_MAX_USER_LINE_LENGTH,
+    ck_assert_msg(strlen(SOPC_Check_Logger_lastUserLog) == (size_t) SOPC_LOG_MAX_USER_LINE_LENGTH,
+                  "Was expecting len=%zu, but found %zu", (size_t) SOPC_LOG_MAX_USER_LINE_LENGTH,
                   strlen(SOPC_Check_Logger_lastUserLog));
 
     // Check sections
